Refactor de archivos de carpeta
